About Us and News Page
Created two new webpages for the project.

The "About Us" webpage has a five paragraph section for details about the company as well as a placeholder area for a 900 x 400 pixel photo.

To the right of the about us description is a heading section that acts as an area to provide additional context as the user reads the about section.

The nav bar header, footer, and footer nav bar are the same as featured on the home page.

The news page features a section to insert four articles with a brief description of them. And the the right of it is an area for news categories and tags.

The nav bar header, footer, and footer nav bar are also the same as featured on the home page.